# backend/api/swagger.yaml
swagger: "2.0"
info:
  title: Accountability Expense-Tracker API
  version: 1-oas3
host: accountability-190955e8b06f.herokuapp.com
basePath: /api
schemes:
  - https
paths:
...
  /trips/create:
    post:
      summary: CREATE trip endpoint
      description: Create a new, empty trip
      parameters:
      - in: body
        name: JSON
        schema:
          type: object
          properties:
            leaderId:
              type: string
              description: ID of the user that owns the trip
            name:
              type: string
              description: Title of the trip
            notes:
              type: string
              description: Miscellaneous description for the trip
      responses:
        "200":
          description: OK
          schema:
            type: object
            properties:
              tripId:
                description: The ID of the created trip
              token:
                type: string
                description: Refreshed JWT
        "400":
          description: Malformed request or JSON
          schema:
            $ref: '#/responses/BadRequest'
  /trips/get:
    post:
      summary: READ trip endpoint
      description: Get a trip with its associated users and expenses
      parameters:
      - in: body
        name: JSON
        schema:
          type: object
          properties:
            tripId:
              type: string
              description: ID of the trip to read
      responses:
        "200":
          description: OK
          schema:
            type: object
            properties:
              token:
                type: string
                description: Refreshed JWT
              leaderId:
                type: string
                description: ID of the user that owns the trip
              name:
                type: string
                description: Title of the trip
              notes:
                type: string
                description: Miscellaneous description for the trip 
              inviteCode:
                type: string
                description: 6-digit invite code
              allMembers:
                type: array
                items:
                  type: object
                  properties:
                    _id:
                      type: string
                    name:
                      type: string
                    email:
                      type: string
              allExpenses:
                type: array
                items:
                  type: object
                  properties:
                    _id:
                      type: string
                    name:
                      type: string
                    description:
                      type: string
                    cost:
                      type: number
        "400":
          description: Malformed request or JSON
          schema:
            $ref: '#/responses/BadRequest'
  /trips/update:
    post:
      summary: UPDATE trip endpoint
      description: Modify trip information
      parameters:
      - in: body
        name: JSON
        schema:
          type: object
          properties:
            tripId:
              type: string
              description: ID of the trip to modify
      responses:
        "200":
          description: OK
          schema:
            type: object
            properties:
              tripId:
                description: Same as input ID
              token:
                type: string
                description: Refreshed JWT
        "400":
          description: Malformed request or JSON
          schema:
            $ref: '#/responses/BadRequest'
  /trips/delete:
    post:
      summary: DELETE trip endpoint
      description: Remove a trip and associated data
      parameters:
      - in: body
        name: JSON
        schema:
          type: object
          properties:
            tripId:
              type: string
              description: ID of the trip to delete
      responses:
        "200":
          description: OK
          schema:
            type: object
            properties:
              token:
                type: string
                description: Refreshed JWT
        "400":
          description: Malformed request or JSON
          schema:
            $ref: '#/responses/BadRequest'
  /trips/listMemberOf:
    post:
      summary: List trips as member user
      description: List trips a user is part of, only shows member (non-leader) trips.
      parameters:
      - in: body
        name: JSON
        schema:
          type: object
          properties:
            userId:
              type: string
              description: ID of the user to look for
      responses:
        "200":
          description: OK
          schema:
            type: object
            properties:
              token:
                type: string
                description: Refreshed JWT
              trips:
                type: array
                items:
                  type: object
                  properties:
                    _id:
                      type: string
                      description: ID of the trip, suitable for other /trips/ endpoint calls
                    name:
                      type: string
                      description: Title of the trip
                    notes:
                      type: string
                      description: Miscellaneous description for the trip
        "400":
          description: Malformed request or JSON
          schema:
            $ref: '#/responses/BadRequest'
  /trips/listOwnerOf:
    post:
      summary: List trips owned by user
      description: List trips that belong to a user, only shows owned trips where the user is the Leader.
      parameters:
      - in: body
        name: JSON
        schema:
          type: object
          properties:
            userId:
              type: string
              description: ID of the user to look for
      responses:
        "200":
          description: OK
          schema:
            type: object
            properties:
              token:
                type: string
                description: Refreshed JWT
              trips:
                type: array
                items:
                  type: object
                  properties:
                    _id:
                      type: string
                      description: ID of the trip, suitable for other /trips/ endpoint calls
                    name:
                      type: string
                      description: Title of the trip
                    notes:
                      type: string
                      description: Miscellaneous description for the trip
        "400":
          description: Malformed request or JSON
          schema:
            $ref: '#/responses/BadRequest'
...
definitions:
...
    ErrorInfo:
      type: object
      properties:
        error:
          type: string
          description: The text description of the error.
          example: Information about the error
responses:
  BadRequest:
    description: Malformed request or JSON on front-end
    schema:
      $ref: '#/definitions/ErrorInfo'
...